test: move bear-add-text tests to dedicated add-text.test.ts
Extracts append/prepend regression tests from replace-text.test.ts into
their own file so bear-add-text tests are discoverable by file name.
Test names updated to describe behavior rather than regression context.

Created with Claude Code under the supervision of Serhii Vasylenko
