+++ tests/system/replace-text.test.ts
@@ -273,70 +273,4 @@ describe('bear-replace-text via MCP Inspector CLI', () => {
       if (noteId) archiveNote(noteId);
     }
   });
-
-  describe('bear-add-text regression', () => {
-    it('regression — prepend still works', () => {
-      const title = uniqueTitle(TEST_PREFIX, 'Prepend Default', RUN_ID);
-      let noteId: string | undefined;
-
-      try {
-        callTool({
-          toolName: 'bear-create-note',
-          args: { title, text: 'Original content', tags: 'system-test' },
-        });
-
-        noteId = findNoteId(title);
-
-        callTool({
-          toolName: 'bear-add-text',
-          args: { id: noteId, text: 'Prepended text', position: 'beginning' },
-        });
-
-        syncSleep(500);
-
-        const openResult = callTool({
-          toolName: 'bear-open-note',
-          args: { id: noteId },
-        });
-
-        const noteBody = extractNoteBody(openResult);
-        expect(noteBody).toContain('Original content');
-        expect(noteBody).toContain('Prepended text');
-      } finally {
-        if (noteId) archiveNote(noteId);
-      }
-    });
-
-    it('regression — default append still works', () => {
-      const title = uniqueTitle(TEST_PREFIX, 'Append Default', RUN_ID);
-      let noteId: string | undefined;
-
-      try {
-        callTool({
-          toolName: 'bear-create-note',
-          args: { title, text: 'Original content', tags: 'system-test' },
-        });
-
-        noteId = findNoteId(title);
-
-        callTool({
-          toolName: 'bear-add-text',
-          args: { id: noteId, text: 'Appended text' },
-        });
-
-        syncSleep(500);
-
-        const openResult = callTool({
-          toolName: 'bear-open-note',
-          args: { id: noteId },
-        });
-
-        const noteBody = extractNoteBody(openResult);
-        expect(noteBody).toContain('Original content');
-        expect(noteBody).toContain('Appended text');
-      } finally {
-        if (noteId) archiveNote(noteId);
-      }
-    });
-  });
 });
